texture inspector class cast exception resolved
git-svn-id: http://fs.math.tu-berlin.de:8000/svn/jreality/trunk@5267 6a571d9c-c4e9-67b8-d2a0-cd0d85a3e679

+++ src-plugin/de/jreality/ui/TextureInspector.java
@@ -82,11 +82,11 @@ public class TextureInspector extends JPanel implements ChangeListener {
 		scaleUSlider = new JSliderVR(SwingConstants.HORIZONTAL, 0, 100, 0),
 		scaleVSlider = new JSliderVR(SwingConstants.HORIZONTAL, 0, 100, 0);
 	private SpinnerNumberModel
-		rotateModel = new SpinnerNumberModel(0.0, -180, 180, 0.1),
-		translateUModel = new SpinnerNumberModel(0.0, 0, 1000.0, 0.001),
-		translateVModel = new SpinnerNumberModel(0.0, 0, 1000.0, 0.001),
-		scaleUModel = new SpinnerNumberModel(1.0, 0, 1000.0, 0.001),
-		scaleVModel = new SpinnerNumberModel(1.0, 0, 1000.0, 0.001);
+		rotateModel = new SpinnerNumberModel(0.0, -180.0, 180.0, 0.1),
+		translateUModel = new SpinnerNumberModel(0.0, 0.0, 1000.0, 0.001),
+		translateVModel = new SpinnerNumberModel(0.0, 0.0, 1000.0, 0.001),
+		scaleUModel = new SpinnerNumberModel(1.0, 0.0, 1000.0, 0.001),
+		scaleVModel = new SpinnerNumberModel(1.0, 0.0, 1000.0, 0.001);
 	private JSpinner
 		rotateSpinner = new JSpinner(rotateModel),
 		translateUSpinner = new JSpinner(translateUModel),
@@ -363,7 +363,7 @@ public class TextureInspector extends JPanel implements ChangeListener {
 		}
 		if (rotateSlider == s) {
 			rotateSlider.removeChangeListener(this);
-			rotateModel.setValue(rotateSlider.getValue());
+			rotateModel.setValue((double)rotateSlider.getValue());
 			rotateSlider.addChangeListener(this);
 		}
 		if (rotateSpinner == s) {
